Add blog entry for authorship of strava-runmap
This commit adds the blog entry detailing the
creation of strava-runmap. It is a short entry
intended to explain the page as it is on my site,
instruct anybody who might want to use it how to,
and give proper credit to the parties and tools
that inspired it and made it possible.